third_party/blink/web_tests/external/wpt/html/semantics/forms/historical-expected.txt

This is a testharness.js-based test.
[FAIL] window.onsearch should not be supported
  assert_false: expected false got true
[FAIL] document.onsearch should not be supported
  assert_false: expected false got true
[FAIL] input.onsearch should not be supported
  assert_false: expected false got true
[FAIL] input.incremental should not be supported
  assert_false: expected false got true
Harness: the test ran to completion.
